# openapi.yaml
openapi: 3.1.0
info:
  title: Pong API
  description: |
    Pong is a health monitoring API that lets you track uptime, latency, and
    availability of your services. Create monitors for HTTP endpoints, SSL
    certificates, heartbeat checks, keyword presence, and port
    availability. Configure alert contacts to get notified when things go wrong,
    and publish public status pages so your users stay informed.

    ## Authentication

    All endpoints under `/api/v1/` require authentication via one of:

    - **Auth0 JWT** — RS256 Bearer token issued by Auth0 (used by the web app)
    - **API Key** — Long-lived key in the format `pong_<hex>` (used for scripts and CI/CD)

    Both are passed as `Authorization: Bearer <token>`.

    API keys can be created and managed via the `/api/v1/api-keys` endpoints.
    The raw key is only shown once at creation time.

    Public endpoints such as heartbeat pings and status pages do not require
    authentication.
  version: 1.0.0
servers:
  - url: https://api.getpong.dev
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 RS256 JSON Web Token
...
  parameters:
    MonitorId:
      name: id
      in: path
      required: true
      description: Unique monitor identifier.
      schema:
        type: integer
        example: 42
...
  responses:
    Unauthorized:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "unauthorized"
...
    BadRequest:
      description: The request body is invalid or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid request body"
paths:
...
  /api/v1/monitors:
    get:
      operationId: listMonitors
      tags: [Monitors]
      summary: List monitors
      description: Returns all monitors belonging to the authenticated user.
      responses:
        "200":
          description: A list of monitors.
          content:
            application/json:
              schema:
                type: object
                required: [monitors]
                properties:
                  monitors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Monitor"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createMonitor
      tags: [Monitors]
      summary: Create a monitor
      description: |
        Creates a new monitor for the authenticated user.

        For **heartbeat** monitors the `target` field is not required; the
        backend will generate a unique `heartbeat_token` that can be used to
        ping the heartbeat endpoint.

        Default values when not specified:
        - `interval_secs`: 300
        - `timeout_secs`: 30
        - `expected_status`: 200
        - `confirmation_count`: 1
        - `ssl_warn_days`: 30
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorCreate"
            examples:
              http_monitor:
                summary: HTTP monitor
                value:
                  name: "Production API"
                  type: "http"
                  target: "https://api.example.com/healthz"
                  interval_secs: 60
                  expected_status: 200
                  alert_contact_ids: [1]
              heartbeat_monitor:
                summary: Heartbeat monitor
                value:
                  name: "Cron Job Heartbeat"
                  type: "heartbeat"
                  interval_secs: 3600
              keyword_monitor:
                summary: Keyword monitor
                value:
                  name: "Homepage keyword check"
                  type: "keyword"
                  target: "https://example.com"
                  keyword: "Welcome"
                  keyword_type: "contains"
      responses:
        "201":
          description: Monitor created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/monitors/{id}:
    get:
      operationId: getMonitor
      tags: [Monitors]
      summary: Get a monitor
      description: Returns a single monitor by ID.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      responses:
        "200":
          description: The requested monitor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateMonitor
      tags: [Monitors]
      summary: Update a monitor
      description: Updates an existing monitor. Only the fields provided in the request body are changed.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorUpdate"
      responses:
        "200":
          description: Monitor updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteMonitor
      tags: [Monitors]
      summary: Delete a monitor
      description: Permanently deletes a monitor and all its associated check results.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      responses:
        "204":
          description: Monitor deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/monitors/{id}/pause:
    post:
      operationId: pauseMonitor
      tags: [Monitors]
      summary: Pause a monitor
      description: Pauses an active monitor so it stops performing checks.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      responses:
        "200":
          description: Monitor paused.
          content:
            application/json:
              schema:
                type: object
                required: [enabled]
                properties:
                  enabled:
                    type: boolean
                    example: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/monitors/{id}/resume:
    post:
      operationId: resumeMonitor
      tags: [Monitors]
      summary: Resume a monitor
      description: Resumes a paused monitor so it starts performing checks again.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      responses:
        "200":
          description: Monitor resumed.
          content:
            application/json:
              schema:
                type: object
                required: [enabled]
                properties:
                  enabled:
                    type: boolean
                    example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/monitors/{id}/results:
    get:
      operationId: listCheckResults
      tags: [Monitors]
      summary: List check results
      description: Returns paginated check results for a monitor, ordered by most recent first.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
        - name: limit
          in: query
          description: Maximum number of results to return.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
            example: 50
        - name: offset
          in: query
          description: Number of results to skip for pagination.
          schema:
            type: integer
            default: 0
            minimum: 0
            example: 0
      responses:
        "200":
          description: A paginated list of check results.
          content:
            application/json:
              schema:
                type: object
                required: [results, total, limit, offset]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/CheckResult"
                  total:
                    type: integer
                    description: Total number of results available.
                    example: 1440
                  limit:
                    type: integer
                    description: Maximum results returned per page.
                    example: 50
                  offset:
                    type: integer
                    description: Number of results skipped.
                    example: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/monitors/{id}/uptime:
    get:
      operationId: getUptime
      tags: [Monitors]
      summary: Get uptime percentage
      description: Returns the uptime percentage for a monitor over a specified number of hours.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
        - name: hours
          in: query
          description: Number of hours to calculate uptime over.
          schema:
            type: integer
            default: 24
            minimum: 1
            example: 24
      responses:
        "200":
          description: Uptime percentage for the specified period.
          content:
            application/json:
              schema:
                type: object
                required: [monitor_id, hours, percentage]
                properties:
                  monitor_id:
                    type: integer
                    description: ID of the monitor.
                    example: 42
                  hours:
                    type: integer
                    description: Number of hours covered.
                    example: 24
                  percentage:
                    type: number
                    format: double
                    description: Uptime percentage (0-100).
                    example: 99.98
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/monitors/{id}/uptime/daily:
    get:
      operationId: getDailyUptime
      tags: [Monitors]
      summary: Get daily uptime
      description: Returns day-by-day uptime data for a monitor over a specified number of days.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
        - name: days
          in: query
          description: Number of days of daily uptime data to return.
          schema:
            type: integer
            default: 90
            minimum: 1
            example: 90
      responses:
        "200":
          description: Daily uptime breakdown.
          content:
            application/json:
              schema:
                type: object
                required: [monitor_id, days, daily]
                properties:
                  monitor_id:
                    type: integer
                    description: ID of the monitor.
                    example: 42
                  days:
                    type: integer
                    description: Number of days covered.
                    example: 90
                  daily:
                    type: array
                    items:
                      $ref: "#/components/schemas/DailyUptime"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Alert Contacts ----------
  /api/v1/alert-contacts:
    get:
      operationId: listAlertContacts
      tags: [Alert Contacts]
      summary: List alert contacts
      description: Returns all alert contacts belonging to the authenticated user.
      responses:
        "200":
          description: A list of alert contacts.
          content:
            application/json:
              schema:
                type: object
                required: [contacts]
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/AlertContact"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createAlertContact
      tags: [Alert Contacts]
      summary: Create an alert contact
      description: Creates a new alert contact for the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertContactCreate"
            examples:
              email:
                summary: Email contact
                value:
                  name: "Ops Team"
                  type: "email"
                  value: "ops@example.com"
              slack:
                summary: Slack webhook
                value:
                  name: "Engineering Slack"
                  type: "slack"
                  value: "https://hooks.slack.com/services/T00/B00/xxx"
              webhook:
                summary: Generic webhook
                value:
                  name: "PagerDuty"
                  type: "webhook"
                  value: "https://events.pagerduty.com/integration/abc/enqueue"
      responses:
        "201":
          description: Alert contact created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertContact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/alert-contacts/{id}:
    put:
      operationId: updateAlertContact
      tags: [Alert Contacts]
      summary: Update an alert contact
      description: Updates an existing alert contact. Only the fields provided in the request body are changed.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertContactUpdate"
      responses:
        "200":
          description: Alert contact updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertContact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteAlertContact
      tags: [Alert Contacts]
      summary: Delete an alert contact
      description: Permanently deletes an alert contact.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      responses:
        "204":
          description: Alert contact deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alert-contacts/{id}/test:
    post:
      operationId: testAlertContact
      tags: [Alert Contacts]
      summary: Send test alert
      description: |
        Sends a test alert notification to the specified contact. Useful for
        verifying that email, Slack, or webhook integrations are configured
        correctly.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      responses:
        "200":
          description: Test alert sent successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "sent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Test alert delivery failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/alert-contacts/{id}/resend:
    post:
      operationId: resendVerification
      tags: [Alert Contacts]
      summary: Resend verification email
      description: |
        Resends the verification email for an unverified email contact.
        Returns 404 if the contact is already verified or doesn't exist.
      parameters:
        - $ref: "#/components/parameters/AlertContactId"
      responses:
        "200":
          description: Verification email sent.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "sent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /verify-email/{token}:
    get:
      operationId: verifyEmail
      tags: [Alert Contacts]
      summary: Verify email address
      description: |
        Verifies an email alert contact by token. This is a public endpoint
        accessed by clicking the link in the verification email. Returns an
        HTML confirmation page.
      security: []
      parameters:
        - name: token
          in: path
          required: true
          description: Verification token from the email link.
          schema:
            type: string
      responses:
        "200":
          description: Email verified. Returns HTML confirmation page.
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Invalid or expired verification token.

  # ---------- Status Pages ----------
  /api/v1/status-pages:
    get:
      operationId: listStatusPages
      tags: [Status Pages]
      summary: List status pages
      description: Returns all status pages belonging to the authenticated user.
      responses:
        "200":
          description: A list of status pages.
          content:
            application/json:
              schema:
                type: object
                required: [status_pages]
                properties:
                  status_pages:
                    type: array
                    items:
                      $ref: "#/components/schemas/StatusPage"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createStatusPage
      tags: [Status Pages]
      summary: Create a status page
      description: |
        Creates a new status page for the authenticated user. A unique token is
        generated automatically and used as the public URL path. An optional
        password can be set to restrict access.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusPageCreate"
      responses:
        "201":
          description: Status page created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/status-pages/{id}:
    get:
      operationId: getStatusPage
      tags: [Status Pages]
      summary: Get a status page
      description: Returns a single status page by ID.
      parameters:
        - $ref: "#/components/parameters/StatusPageId"
      responses:
        "200":
          description: The requested status page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateStatusPage
      tags: [Status Pages]
      summary: Update a status page
      description: |
        Updates an existing status page. Only the fields provided in the request
        body are changed. To set a password, include the password field. To remove
        a password, send an empty string. Omit the field to keep it unchanged.
      parameters:
        - $ref: "#/components/parameters/StatusPageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusPageUpdate"
      responses:
        "200":
          description: Status page updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteStatusPage
      tags: [Status Pages]
      summary: Delete a status page
      description: Permanently deletes a status page.
      parameters:
        - $ref: "#/components/parameters/StatusPageId"
      responses:
        "204":
          description: Status page deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Heartbeat (Public) ----------
  # ---------- API Keys ----------
  /api/v1/api-keys:
    get:
      operationId: listAPIKeys
      tags: [API Keys]
      summary: List API keys
      description: Returns all API keys for the authenticated user. Raw keys are never returned.
      responses:
        "200":
          description: List of API keys.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: "#/components/schemas/APIKey"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createAPIKey
      tags: [API Keys]
      summary: Create API key
      description: |
        Generates a new API key. The raw key is returned in the response and will
        **never be shown again**. Store it securely.

        Use the key as `Authorization: Bearer pong_...` in API requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: A descriptive name for the key.
                  example: "CI Pipeline"
      responses:
        "201":
          description: API key created. The raw key is included in this response only.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIKeyCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/api-keys/{id}:
    delete:
      operationId: deleteAPIKey
      tags: [API Keys]
      summary: Delete API key
      description: Permanently revokes an API key. Any requests using this key will immediately fail.
      parameters:
        - $ref: "#/components/parameters/APIKeyId"
      responses:
        "204":
          description: API key deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Heartbeat ----------
  /api/v1/heartbeat/{token}:
    get:
      operationId: pingHeartbeatGet
      tags: [Heartbeat]
      summary: Ping a heartbeat monitor (GET)
      description: |
        Records a heartbeat ping for the monitor associated with the given token.
        This endpoint does not require authentication and is intended to be called
        from cron jobs, scripts, or other automated processes.

        If the monitor has a `heartbeat_secret` configured, the request must include
        an `X-Secret` header with the matching value, otherwise returns 401.
      security: []
      parameters:
        - $ref: "#/components/parameters/HeartbeatToken"
        - name: X-Secret
          in: header
          required: false
          description: Required if the heartbeat monitor has a secret configured.
          schema:
            type: string
      responses:
        "200":
          description: Heartbeat recorded successfully.
          content:
            application/json:
              schema:
                type: object
                required: [ok, monitor_id]
                properties:
                  ok:
                    type: boolean
                    example: true
                  monitor_id:
                    type: integer
                    example: 42
        "401":
          description: Secret mismatch. The X-Secret header is missing or does not match.
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: pingHeartbeatPost
      tags: [Heartbeat]
      summary: Ping a heartbeat monitor (POST)
      description: |
        Records a heartbeat ping for the monitor associated with the given token.
        Behaves identically to the GET variant. This endpoint does not require
        authentication.

        If the monitor has a `heartbeat_secret` configured, the request must include
        an `X-Secret` header with the matching value, otherwise returns 401.
      security: []
      parameters:
        - $ref: "#/components/parameters/HeartbeatToken"
        - name: X-Secret
          in: header
          required: false
          description: Required if the heartbeat monitor has a secret configured.
          schema:
            type: string
      responses:
        "200":
          description: Heartbeat recorded successfully.
          content:
            application/json:
              schema:
                type: object
                required: [ok, monitor_id]
                properties:
                  ok:
                    type: boolean
                    example: true
                  monitor_id:
                    type: integer
                    example: 42
        "401":
          description: Secret mismatch. The X-Secret header is missing or does not match.
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Public Status Page ----------
  /status/{token}:
    get:
      operationId: getPublicStatusPage
      tags: [Status Pages]
      summary: View public status page
      description: |
        Returns an HTML status page for public viewing. If the page is password
        protected, returns a password form instead. No authentication required.
      security: []
      parameters:
        - $ref: "#/components/parameters/StatusPageToken"
      responses:
        "200":
          description: HTML status page or password form.
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Status page not found.

    post:
      operationId: submitStatusPagePassword
      tags: [Status Pages]
      summary: Submit password for protected status page
      description: |
        Submits the password for a protected status page. On success, sets a
        session cookie and redirects to the GET endpoint. On failure, re-renders
        the password form with an error message.
      security: []
      parameters:
        - $ref: "#/components/parameters/StatusPageToken"
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                password:
                  type: string
      responses:
        "303":
          description: Password correct. Redirects to the status page with a session cookie.
        "200":
          description: Wrong password. Re-renders the password form with an error.
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Status page not found.

  # ---------- Admin ----------
  /api/v1/admin/stats:
    get:
      operationId: getAdminStats
      tags: [Admin]
      summary: Platform statistics
      description: Returns total users, monitors, active monitors, and waitlist count. Requires admin role.
      responses:
        "200":
          description: Platform stats.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_users:
                    type: integer
                  total_monitors:
                    type: integer
                  active_monitors:
                    type: integer
                  waitlist_count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Not found (non-admin users receive 404).
...
  /api/v1/admin/users/{id}/plan:
    put:
      operationId: setUserPlan
      tags: [Admin]
      summary: Change user plan
      description: Updates a user's subscription plan. Requires admin role.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan]
              properties:
                plan:
                  type: string
                  enum: [free, pro, business, selfhosted]
      responses:
        "200":
          description: Plan updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...